requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
gunicorn==21.2.0
django-celery-results==2.5.1